book: Split into bin and lib target
Enables more flexible importing of components into other crates,
which is also useful for testing.
