+++ book/src/main.rs
@@ -1,28 +1,23 @@
-use config::Configuration;
 use diesel::prelude::*;
 use diesel::r2d2::{ConnectionManager, Pool};
 use dotenv::dotenv;
 use envconfig::Envconfig;
-use futures::{future, prelude::*};
 use once_cell::sync::OnceCell;
-use rpc::server::BookServer;
-use rpc::service::BookService;
 use std::io;
 use std::process;
-use tarpc::server::{self, Channel, Handler};
-use tokio_serde::formats::Json;
+
+use book_lib::{rpc_server, DB};
 
 #[macro_use]
 extern crate diesel_migrations;
 
 mod config;
-mod db;
-mod rpc;
+
+use crate::config::Configuration;
 
 static PKG_NAME: Option<&'static str> = option_env!("CARGO_PKG_NAME");
 static PKG_VERSION: Option<&'static str> = option_env!("CARGO_PKG_VERSION");
 static CONFIGURATION: OnceCell<Configuration> = OnceCell::new();
-static DB: OnceCell<Pool<ConnectionManager<PgConnection>>> = OnceCell::new();
 
 #[tokio::main]
 async fn main() -> io::Result<()> {
@@ -78,18 +73,11 @@ async fn main() -> io::Result<()> {
     embed_migrations!();
     helpers::db::run_migration(embedded_migrations::run, &DB.get().unwrap());
 
-    tarpc::serde_transport::tcp::listen(&CONFIGURATION.get().unwrap().rpc_socket(), Json::default)
-        .await?
-        .filter_map(|r| future::ready(r.ok()))
-        .map(server::BaseChannel::with_defaults)
-        .max_channels_per_key(1, |t| t.as_ref().peer_addr().unwrap().ip())
-        .map(|channel| {
-            let server = BookServer(channel.as_ref().as_ref().peer_addr().unwrap());
-            channel.respond_with(server.serve()).execute()
-        })
-        .buffer_unordered(10)
-        .for_each(|_| async {})
-        .await;
+    let (server, addr) = rpc_server(&CONFIGURATION.get().unwrap().rpc_socket())
+        .await
+        .unwrap();
+    log::error!("Book RPC Server started on {}", addr);
+    server.await;
 
     Ok(())
 }
